Delete and create new table Oppfolgingstilfelle
Delete table Oppfolgingstilfelle that is unused in production. Oppfolgingstilfelle will be based on the event Oversikthendelsetilfelle instead of Oppfolgingstilfelle. The data is therefore deleted and a new table with fnr instead of aktorId is created to store data from Oversikthendelsetilfelle. This change will ensure data consistency with the database in Syfooversikt and its Person-Oppfolgingstilfelle list.