Add Motebehov in entire Oppfolgingstilfelle
Multiple active Oppfolgingstilfeller are merged into one for both Arbeidsgiver and Arbeidstaker. Endpoints and tests for Arbeidstaker and arbeidstaker are split into separate files. Functionality for Motebehov available in entire Oppfolgingstilfelle is added and Motebehov is split into two types: MELD_BEHOV and SVAR_BEHOV. MELD_BEHOV is in the intervals [week 0, week 16) and (week 26, end of Oppfolgingstilfelle], while SVAR_BEHOV is available in the interval [16, 26)