GET existing motebehov with new motebehovSvar fields (#405)
Add new fields to the "output DTO" for the motebehovSvar part when GETting motebehov from controllers.

Before, motebehov only has harMotebehov and forklaring as part of their motebehovSvar output DTO. Now, they will also have fields begrunnelse, onskerSykmelderDeltar, onskerSykmelderDeltarBegrunnelse and more. These fields correspond to the new fields that are to be added to the forms in our frontend. Most of these will be null for existing motebehov. But begrunnelse and onskerSykmelderDeltar might be populated, depending on the contents of forklaring of an existing motebehov.

If a forklaring of an existing motebehov contained "Jeg ønsker at den som sykmelder arbeidstakeren deltar i møtet. Begrunnelsestekst", then the new begrunnelse field will only contain "Begrunnelsestekst" and onskerSykmelderDeltar will be true. This is to "revert" a hack we did in the frontend where we "combined" two form fields, and sent is as one string (forklaring) to backend and stored it as one string in database.

Old motebehov svar will also be converted to the new formFillout field in the motebehovSvar output DTO. The formFillout value can be used to display what the form looked like when it was filled out, including details like field labels and values and unselected options.

Also
* Use NAV Maven repository for github packages
* Don't mark MotebehovSvar classes as Serializable
---------

Co-authored-by: malazkoj <malaz.alkoj@nav.no>